odst_tool/requirements.txt
spotipy>=2.23.0
yt-dlp>=2023.0.0
mutagen>=1.47.0
rich>=13.0.0
python-dotenv>=1.0.0
requests>=2.28.0
pillow>=10.0.0
flask>=3.0.0
flask-socketio>=5.0.0
eventlet>=0.30.0
boto3>=1.26.0
